Remove BAR API client factory
