+++ src/Microsoft.DotNet.Darc/Darc/Operations/AddChannelOperation.cs
@@ -31,7 +31,7 @@ internal class AddChannelOperation : Operation
     {
         try
         {
-            IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+            IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
             // If the user tried to mark as internal, indicate that this is currently
             // unsupported.
