+++ src/Microsoft.DotNet.Darc/Darc/Operations/AddSubscriptionOperation.cs
@@ -35,7 +35,7 @@ class AddSubscriptionOperation : Operation
     /// <param name="options"></param>
     public override async Task<int> ExecuteAsync()
     {
-        IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+        IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
         if (_options.IgnoreChecks.Any() && !_options.AllChecksSuccessfulMergePolicy)
         {
