+++ src/Microsoft.DotNet.Darc/Darc/Operations/DefaultChannelStatusOperation.cs
@@ -34,7 +34,7 @@ class DefaultChannelStatusOperation : UpdateDefaultChannelBaseOperation
             return Constants.ErrorCode;
         }
 
-        IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+        IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
         try
         {
