+++ src/Microsoft.DotNet.Darc/Darc/Operations/GatherDropOperation.cs
@@ -209,7 +209,7 @@ internal class GatherDropOperation : Operation
             return null;
         }
 
-        IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+        IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
         string repoUri = _options.RepoUri;
 
@@ -689,7 +689,6 @@ internal class GatherDropOperation : Operation
         }
 
         var remoteFactory = new RemoteFactory(_options);
-        var barClientFactory = new BarApiClientFactory(_options);
 
         // Flatten for convenience and remove dependencies of types that we don't want if need be.
         if (!_options.IncludeToolset)
@@ -712,7 +711,7 @@ internal class GatherDropOperation : Operation
             var rootBuildRepository = rootBuild.GitHubRepository ?? rootBuild.AzureDevOpsRepository;
             DependencyGraph graph = await DependencyGraph.BuildRemoteDependencyGraphAsync(
                 remoteFactory,
-                await barClientFactory.GetBasicBarClient(Logger),
+                RemoteFactory.GetBarClient(_options, Logger),
                 rootBuildRepository,
                 rootBuild.Commit,
                 buildOptions,
@@ -804,7 +803,7 @@ internal class GatherDropOperation : Operation
     /// <param name="rootOutputDirectory">Output directory. Must exist.</param>
     private async Task<DownloadedBuild> GatherDropForBuildAsync(Build build, string rootOutputDirectory)
     {
-        IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+        IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
         var success = true;
         var unifiedOutputDirectory = rootOutputDirectory;
         Directory.CreateDirectory(unifiedOutputDirectory);
