+++ src/Microsoft.DotNet.Darc/Darc/Operations/GetAssetOperation.cs
@@ -37,7 +37,7 @@ internal class GetAssetOperation : Operation
             return Constants.ErrorCode;
         }
 
-        IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+        IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
         try
         {
