+++ src/Microsoft.DotNet.Darc/Darc/Operations/GetBuildOperation.cs
@@ -37,7 +37,7 @@ internal class GetBuildOperation : Operation
         try
         {
             IBarApiClient remote = Provider.GetService<IBarApiClient>()
-                ?? BarApiClientFactory.GetBarClient(_options, Logger);
+                ?? RemoteFactory.GetBarClient(_options, Logger);
 
             List<Build>? matchingBuilds = null;
             if (_options.Id != 0)
