+++ src/Microsoft.DotNet.Darc/Darc/Operations/GetChannelsOperation.cs
@@ -33,7 +33,7 @@ internal class GetChannelsOperation : Operation
     {
         try
         {
-            IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+            IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
             var allChannels = await barClient.GetChannelsAsync();
             switch (_options.OutputFormat)
