+++ src/Microsoft.DotNet.Darc/Darc/Operations/GetHealthOperation.cs
@@ -54,7 +54,7 @@ internal class GetHealthOperation : Operation
     {
         try
         {
-            IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+            IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
             IEnumerable<Subscription> subscriptions = await barClient.GetSubscriptionsAsync();
             IEnumerable<DefaultChannel> defaultChannels = await barClient.GetDefaultChannelsAsync();
@@ -99,14 +99,14 @@ internal class GetHealthOperation : Operation
 
             // Compute metrics, then run in parallel.
 
-            List<Func<Task<HealthMetricWithOutput>>> metricsToRun = await ComputeMetricsToRun(
+            List<Func<Task<HealthMetricWithOutput>>> metricsToRun = ComputeMetricsToRun(
                 channelsToEvaluate,
                 reposToEvaluate,
                 subscriptions,
                 defaultChannels);
 
             // Run the metrics
-            HealthMetricWithOutput[] results = await Task.WhenAll<HealthMetricWithOutput>(metricsToRun.Select(metric => metric()));
+            HealthMetricWithOutput[] results = await Task.WhenAll(metricsToRun.Select(metric => metric()));
 
             // Walk through and print the results out
             bool passed = true;
@@ -140,19 +140,15 @@ internal class GetHealthOperation : Operation
     /// <param name="channelsToEvaluate">Channels to evaluate</param>
     /// <param name="reposToEvaluate">Repositories to evaluate</param>
     /// <returns>List of Func's that, when evaluated, will produce metrics with output.</returns>
-    private async Task<List<Func<Task<HealthMetricWithOutput>>>> ComputeMetricsToRun(
+    private List<Func<Task<HealthMetricWithOutput>>> ComputeMetricsToRun(
         HashSet<string> channelsToEvaluate,
         HashSet<string> reposToEvaluate,
         IEnumerable<Subscription> subscriptions,
-        IEnumerable<DefaultChannel> defaultChannels)
-    {
-        var metricsToRun = new List<Func<Task<HealthMetricWithOutput>>>();
-
-        metricsToRun.AddRange(await ComputeSubscriptionHealthMetricsToRun(channelsToEvaluate, reposToEvaluate, subscriptions, defaultChannels));
-        metricsToRun.AddRange(await ComputeProductDependencyCycleMetricsToRun(channelsToEvaluate, reposToEvaluate, subscriptions, defaultChannels));
-
-        return metricsToRun;
-    }
+        IEnumerable<DefaultChannel> defaultChannels) =>
+    [
+        .. ComputeSubscriptionHealthMetricsToRun(channelsToEvaluate, reposToEvaluate, subscriptions, defaultChannels),
+        .. ComputeProductDependencyCycleMetricsToRun(channelsToEvaluate, reposToEvaluate, subscriptions, defaultChannels),
+    ];
 
     /// <summary>
     ///     Get typical repository and branch combinations for use with repo+branch focused metrics.
@@ -202,15 +198,14 @@ internal class GetHealthOperation : Operation
     ///     Note that this will currently miss completely untargeted branches, until those have at least one
     ///     default channel or subscription. This is a fairly benign limitation.
     /// </remarks>
-    private async Task<List<Func<Task<HealthMetricWithOutput>>>> ComputeSubscriptionHealthMetricsToRun(
+    private List<Func<Task<HealthMetricWithOutput>>> ComputeSubscriptionHealthMetricsToRun(
         HashSet<string> channelsToEvaluate,
         HashSet<string> reposToEvaluate,
         IEnumerable<Subscription> subscriptions,
         IEnumerable<DefaultChannel> defaultChannels)
     {
         var remoteFactory = new RemoteFactory(_options);
-        var barClientFactory = new BarApiClientFactory(_options);
-        var barClient = await barClientFactory.GetBarClientAsync(Logger);
+        var barClient = RemoteFactory.GetBarClient(_options, Logger);
 
         HashSet<(string repo, string branch)> repoBranchCombinations =
             GetRepositoryBranchCombinations(channelsToEvaluate, reposToEvaluate, subscriptions, defaultChannels);
@@ -278,15 +273,14 @@ internal class GetHealthOperation : Operation
     /// <summary>
     ///     Compute product dependency cycle metrics based on the input repositories and channels.
     /// </summary>
-    private async Task<List<Func<Task<HealthMetricWithOutput>>>> ComputeProductDependencyCycleMetricsToRun(
+    private List<Func<Task<HealthMetricWithOutput>>> ComputeProductDependencyCycleMetricsToRun(
         HashSet<string> channelsToEvaluate,
         HashSet<string> reposToEvaluate,
         IEnumerable<Subscription> subscriptions,
         IEnumerable<DefaultChannel> defaultChannels)
     {
         var remoteFactory = new RemoteFactory(_options);
-        var barClientFactory = new BarApiClientFactory(_options);
-        var barClient = await barClientFactory.GetBarClientAsync(Logger);
+        var barClient = RemoteFactory.GetBarClient(_options, Logger);
 
         HashSet<(string repo, string branch)> repoBranchCombinations =
             GetRepositoryBranchCombinations(channelsToEvaluate, reposToEvaluate, subscriptions, defaultChannels);
