+++ src/Microsoft.DotNet.Darc/Darc/Operations/TriggerSubscriptionsOperation.cs
@@ -32,7 +32,7 @@ internal class TriggerSubscriptionsOperation : Operation
     {
         try
         {
-            IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+            IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
             bool noConfirm = _options.NoConfirmation;
             List<Subscription> subscriptionsToTrigger = [];
