+++ src/Microsoft.DotNet.Darc/Darc/Operations/UpdateBuildOperation.cs
@@ -31,7 +31,7 @@ internal class UpdateBuildOperation : Operation
 
         try
         {
-            IBarApiClient barClient = BarApiClientFactory.GetBarClient(_options, Logger);
+            IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
 
             Build updatedBuild = await barClient.UpdateBuildAsync(_options.Id, new BuildUpdate { Released = _options.Released });
