+++ src/Microsoft.DotNet.Darc/Darc/Operations/UpdateDependenciesOperation.cs
@@ -48,8 +48,7 @@ class UpdateDependenciesOperation : Operation
                 _options.GitHubPat;
 
             IRemoteFactory remoteFactory = new RemoteFactory(_options);
-            var barClientFactory = new BarApiClientFactory(_options);
-            IBarApiClient barClient = await barClientFactory.GetBarClientAsync(Logger);
+            IBarApiClient barClient = RemoteFactory.GetBarClient(_options, Logger);
             var coherencyUpdateResolver = new CoherencyUpdateResolver(barClient, Logger);
 
             var local = new Local(_options.GetRemoteConfiguration(), Logger);
